Version 0.3.0 Release
Feature release for Entities [0.17.0]

+++ Core/Core/Systems/DestroyEntitiesOnSceneChangeSystem.cs
@@ -6,7 +6,9 @@ namespace Latios.Systems
 {
     internal struct LatiosSceneChangeDummyTag : IComponentData { }
 
+    [DisableAutoCreation]
     [AlwaysUpdateSystem]
+    [UpdateInGroup(typeof(LatiosInitializationSystemGroup), OrderLast = true)]  //Doesn't matter, but good for visualization
     public class DestroyEntitiesOnSceneChangeSystem : SubSystem
     {
         private EntityQuery m_destroyQuery = default;
@@ -14,7 +16,7 @@ namespace Latios.Systems
         protected override void OnCreate()
         {
             m_destroyQuery =
-                Fluent.WithAll<LatiosSceneChangeDummyTag>().Without<WorldGlobalTag>().Without<DontDestroyOnSceneChangeTag>().IncludePrefabs().IncludeDisabled().Build();
+                Fluent.WithAll<LatiosSceneChangeDummyTag>().Without<WorldBlackboardTag>().Without<DontDestroyOnSceneChangeTag>().IncludePrefabs().IncludeDisabled().Build();
             SceneManager.activeSceneChanged += RealUpdateOnSceneChange;
         }
 
@@ -29,11 +31,14 @@ namespace Latios.Systems
 
         private void RealUpdateOnSceneChange(Scene unloaded, Scene loaded)
         {
+            if (!Enabled)
+                return;
+
             latiosWorld.ResumeNextFrame();
             EntityManager.AddComponent<LatiosSceneChangeDummyTag>(EntityManager.UniversalQuery);
             EntityManager.DestroyEntity(m_destroyQuery);
             EntityManager.RemoveComponent<LatiosSceneChangeDummyTag>(EntityManager.UniversalQuery);
-            latiosWorld.CreateNewSceneGlobalEntity();
+            latiosWorld.CreateNewSceneBlackboardEntity();
         }
     }
 }
